requirements.txt
alembic==1.13.1
APScheduler==3.10.4
async-timeout==4.0.3
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
Flask==3.0.3
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.12.0
lxml==5.2.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysqlclient==2.2.4
ordered-set==4.1.0
packaging==24.1
Pygments==2.18.0
PyPDF2==3.0.1
python-docx==1.1.2
pytz==2024.1
redis==5.0.6
requests==2.32.3
rich==13.7.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.30
stripe==9.12.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
waitress==3.0.0
Werkzeug==3.0.3
wrapt==1.16.0
WTForms==3.1.2
